# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects": {
      "post": {
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "a project with that name already exists (I-PRJ-4)"
          }
        },
        "summary": "Create a project (git ⇒ async clone, empty ⇒ ready); always 202",
        "tags": [
          "project"
        ]
      },
      "get": {
        "operationId": "ProjectController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all projects",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "operationId": "ProjectController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a project by id",
        "tags": [
          "project"
        ]
      },
      "delete": {
        "operationId": "ProjectController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProjectDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a project (optionally keeping its baseline)",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/retry-clone": {
      "post": {
        "operationId": "ProjectController_retryClone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Retry a failed clone",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/convert-to-empty": {
      "post": {
        "operationId": "ProjectController_convertToEmpty",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Convert a failed git project into an empty one",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/cancel-clone": {
      "post": {
        "operationId": "ProjectController_cancelClone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Cancel an in-progress clone",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/branches": {
      "get": {
        "operationId": "ProjectController_listBranches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "summary": "List the branches in the project's baseline (local refs; no network)",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/sync": {
      "post": {
        "operationId": "ProjectController_sync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "the project is not ready (or has no remote)"
          }
        },
        "summary": "Sync the baseline from its remote (git fetch --all)",
        "tags": [
          "project"
        ]
      }
    },
    "/api/projects/{id}/deletion-preview": {
      "get": {
        "operationId": "ProjectController_deletionPreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDeletionPreviewResponseDto"
                }
              }
            }
          }
        },
        "summary": "Read authoritative project deletion blockers and consequence counts",
        "tags": [
          "project"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agent Platform API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "git",
              "empty"
            ]
          },
          "repoUrl": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "repoBranch": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "name",
          "sourceType"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "git",
              "empty"
            ]
          },
          "cloneStatus": {
            "type": "string",
            "enum": [
              "cloning",
              "ready",
              "failed"
            ]
          },
          "cloneErrorCode": {
            "type": "string",
            "enum": [
              "CLONE_FAILED_PERMISSION",
              "CLONE_FAILED_NOT_FOUND",
              "CLONE_FAILED_NETWORK",
              "TIMEOUT",
              "INTERRUPTED",
              "DISK_INSUFFICIENT"
            ],
            "nullable": true
          },
          "taskCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "repoUrl": {
            "type": "string"
          },
          "repoBranch": {
            "type": "string"
          },
          "baselineSizeBytes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "sourceType",
          "cloneStatus",
          "cloneErrorCode",
          "taskCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "DeleteProjectDto": {
        "type": "object",
        "properties": {
          "keepBaseline": {
            "type": "boolean"
          }
        }
      },
      "ProjectDeletionPreviewResponseDto": {
        "type": "object",
        "properties": {
          "activeTasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name"
              ]
            }
          },
          "retainedVolumeCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "automationCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "automationRunCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "taskCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          }
        },
        "required": [
          "activeTasks",
          "retainedVolumeCount",
          "automationCount",
          "automationRunCount",
          "taskCount"
        ]
      },
...
    }
  }
}